AsteroidsPartsFactory, SourceAsteroidData

+++ Assets/Game/Code/Unity/Configs/AsteroidsConfig.cs
@@ -33,7 +33,7 @@
 		public float RandomScale => Random.Range( ScaleMin, ScaleMax );
 		public float RandomSpeed => Random.Range( SpeedMin, SpeedMax );
 		public float RandomRotationSpeed => Random.Range( RotationSpeedMin, RotationSpeedMax );
-		public float RandomDestroyParts => Random.Range( DestroyPartsMin, DestroyPartsMax + 1 );
+		public int RandomDestroyParts => Random.Range( DestroyPartsMin, DestroyPartsMax + 1 );
 		public float RandomPartSpeed => Random.Range( PartSpeedMin, PartSpeedMax );
 	}
 }
